fixed issue with not all timeblocks saving by putting single quotes around attribute selector

+++ script.js
@@ -1,26 +1,38 @@
 
-var allTimeBlocks = ["9", "10", "11", "12", "1", "2", "3", "4", "5"];
-
 
 //Event listener to save event text on click of Save button
 $(".save-event").on("click", function() {
 
     var timeAttribute = $(this).attr("time");
-    var newEventText = $("#event[time="+timeAttribute+"]").val();
+    console.log("time block - "+timeAttribute);
+
+    var newEventText = $("#event[time='"+ timeAttribute +"']").val();
     
     console.log(newEventText);
 
-    var planObj = {};
+    
+    var existingSavedEvents = localStorage.getItem("dayPlans");
 
-    planObj[timeAttribute] = newEventText;
+    if(existingSavedEvents) {
 
-    localStorage.setItem("dayPlans", JSON.stringify(planObj));
+        existingSavedEvents = JSON.parse(existingSavedEvents);
+        existingSavedEvents[timeAttribute] = newEventText;
+
+        localStorage.setItem("dayPlans", JSON.stringify(existingSavedEvents));
+
+    }else {
+        var planObj = {};
+        planObj[timeAttribute] = newEventText;
+
+        localStorage.setItem("dayPlans", JSON.stringify(planObj));
+    }
 
 });
 
 //Gets saved events from local storages and adds it to event text in their respective time blocks
 function loadSavedEvents() {
 
+    var allTimeBlocks = ["9", "10", "11", "12", "1", "2", "3", "4", "5"];
     var savedDayPlans = localStorage.getItem("dayPlans");
 
     console.log("saved events - "+savedDayPlans);
@@ -35,7 +47,7 @@ function loadSavedEvents() {
 
                 //If saved event exists, populate the saved event in its respective time block
                 if(savedDayPlans[timeBlock]) {
-                    $("#event[time="+ timeBlock +"]").val(savedDayPlans[timeBlock]);
+                    $("#event[time='"+ timeBlock +"']").val(savedDayPlans[timeBlock]);
                 }
 
             }
@@ -52,7 +64,13 @@ function updateTimeBlockColors() {
     var dateString = moment().format("dddd, MMMM Do YYYY");
     $(".day-display").text(dateString);
     
+    //grey for time blocks past
 
+
+    //light blue for present time block
+
+
+    //dark blue for future time blocks
 
 }
